adciona controller profissionalSaue

+++ SpringProntoClin/src/main/resources/db/migration/V4__create-table-consulta.sql
@@ -2,8 +2,8 @@ CREATE  SEQUENCE consulta_seq START WITH 1 INCREMENT BY 1;
 
 CREATE TABLE IF NOT EXISTS consulta (
 	idconsulta bigint DEFAULT nextval('consulta_seq') PRIMARY KEY,
-	idpaciente int NOT NULL,
-	idprofissionalsaude int NOT NULL,
+	idpaciente bigint NOT NULL,
+	idprofissionalsaude bigint NOT NULL,
 	nomeprofissionalsaude varchar(100) NOT NULL,
 	nomepaciente varchar(100) NOT NULL,
     nomesocial varchar(100),
